Mapeo bien, falta modificacion de datos de perfiles

+++ src/main/java/Giftify/Giftify/Repositories/PerfilRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PerfilRepository extends JpaRepository<Perfil, Integer> { //Es una interfaz de jpa
+public interface PerfilRepository extends JpaRepository<Perfil, Long> { //Es una interfaz de jpa
 }
